requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
